+++ exporters/stdout/stdouttrace/internal/x/x_test.go
@@ -10,15 +10,18 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestSelfObservability(t *testing.T) {
-	const key = "OTEL_GO_X_SELF_OBSERVABILITY"
-	require.Equal(t, key, SelfObservability.Key())
+func TestObservability(t *testing.T) {
+	const key = "OTEL_GO_X_OBSERVABILITY"
+	require.Contains(t, Observability.Keys(), key)
 
-	t.Run("100", run(setenv(key, "100"), assertDisabled(SelfObservability)))
-	t.Run("true", run(setenv(key, "true"), assertEnabled(SelfObservability, "true")))
-	t.Run("True", run(setenv(key, "True"), assertEnabled(SelfObservability, "True")))
-	t.Run("false", run(setenv(key, "false"), assertDisabled(SelfObservability)))
-	t.Run("empty", run(assertDisabled(SelfObservability)))
+	const altKey = "OTEL_GO_X_SELF_OBSERVABILITY"
+	require.Contains(t, Observability.Keys(), altKey)
+
+	t.Run("100", run(setenv(key, "100"), assertDisabled(Observability)))
+	t.Run("true", run(setenv(key, "true"), assertEnabled(Observability, "true")))
+	t.Run("True", run(setenv(key, "True"), assertEnabled(Observability, "True")))
+	t.Run("false", run(setenv(key, "false"), assertDisabled(Observability)))
+	t.Run("empty", run(assertDisabled(Observability)))
 }
 
 func run(steps ...func(*testing.T)) func(*testing.T) {